Final Submissions Files
The file titled "FINAL_Model Selection Function and Simulation.R" has the model selection function we wrote, as well as a simulation testing this function. The file titled "FINAL_Testing File.R" has a testthat function we wrote to test our function, as well as a line graph showing how RSS decreases as predictors are added.