FINAL project submission
The file titled "FINAL_Model Selection Function and Simulation" has the function we created, along with simulated data. The file titled "FINAL_Testing" has the testing function we created to verify that our function does what it is supposed to. Finally, the file titled "FINAL_Plotting File" our function, simulation, and a plot showing how RSS decreases as the number of predictors increases.